refactor(#1174): extract select_executor async pipeline into execute.rs submodule
Behavior-preserving decomposition (epic #1116 follow-up). The three large
async pipeline methods relocated intact during the #1121 split — `execute`,
`execute_streaming_background`, `execute_sstable_scan` — are extracted verbatim
into a new `select_executor/execute.rs` as an `impl SelectExecutor` continuation,
following the same submodule pattern as the sibling step helpers.

- mod.rs: 2035 -> 1291 src lines (under the ~1500 soft cap)
- execute.rs: 784 src lines (new, under cap)
- Method bodies byte-identical (relocate, not rewrite); test block untouched,
  test counts preserved exactly.
- Visibility: `execute` stays pub (inherent method, path unchanged);
  `execute_streaming_background` / `execute_sstable_scan` private -> pub(super)
  so mod.rs call sites resolve. No public API change beyond module path;
  re-exports and call sites unchanged.
- No new unwrap()/expect() in library code.

Refs #1116, #1135. Closes #1174.
